Sprite design interface:
A sprite takes a pre-loaded texture, a bounding box in pixel coordinates
and a viewport box in pixel coordinates.

Bounds: location + effective size on the screen
Viewport: section of the original texture which has to be rendered

Tileset and animations:

tilesets are performed by a mapping <integer, rectangle> of the texture
which is able to switch to the referred viewport, kind of multi-viewport sprite.
Animations are simplified by the core::evolvable class.

TODO: Confirm interface, complete it with the renderer flip angle.
      Create an helper function in order to simplify loading of the sprite.

Actual loading method:

    - get, create or load a texture
    - assign it to a new constructed sprite

Texture loading: all textures are mapped by a unique string:
    - path in case of loading
    - name in case of creation (creation fails if the name is already in use)